==> scripts/dada2_pipeline/scripts/01_subset_samples.Rmd <==
---
title: "Subset samples"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Get the asv counts and taxanomy and alpha diversity for that specific project samples

```{r}
# load the three tables
library(vdbR)
library(tidyverse)
connect_database('~/dbConfig.txt')
list_table_from_database('alpha')
get_table_from_database('asv_annotation_blast_ag')
get_table_from_database('asv_alpha_diversity_ag')
cts <- get_counts_subset(c('02.tp.72'))
list_table_from_database()
CS1 <- get_counts_subset(c('CS1'))
```

# Wisconsin samples 

```{r}
alpha <- asv_alpha_diversity_ag %>% 
  filter(uploaded_date == '2021-02-03') %>% 
  select(sampleid:shannon)

cts <- asv_counts_ag %>% 
  filter(sampleid %in% alpha$sampleid) %>% 
  select(asv_key:count_total, count_relative)

taxa <- asv_annotation_blast_ag %>% 
  filter(asv_key %in% cts$asv_key) %>% 
  select(asv_key:blast_pass)

list(alpha = alpha,
     counts = cts,
     taxa = taxa) %>% 
  imap(function(.x, .y){
    write_csv(.x, str_glue('../data/MCW_ASV_{.y}.csv'))
  })
```

Use the matching table from Gabe to count how many samples from how many patients 

```{r}
matching <- readxl::read_excel('../data/Set_90_SampleKey_MCW_matching.xlsx')
alpha <- read_csv('../data/MCW_ASV_alpha.csv')

matched <- matching %>% 
  left_join(alpha, by = c('OligoID' = 'oligos_id'))

matched %>% 
  filter(is.na(simpson_reciprocal))

succe <- matched %>% 
  #filter(!is.na(simpson_reciprocal)) %>% 
  mutate(`Sample ID` = case_when(
  `Sample ID` == 'MCW_0022_A' ~ 'MCW_0022-A',
  `Sample ID` == 'MCW_0021_A' ~ 'MCW_0021-A',
  TRUE ~ `Sample ID`
)) %>% 
  separate(`Sample ID`, into = c('Patient','ID'), sep = '-', remove = F)

succe %>% 
  count(Patient)

succe %>% 
  select(-`Date Collected`) %>% 
  write_csv('../data/MCW_sampleid_patient_match.csv')
```

A total of 122 samples, 15 failed at the lib prep. 107 succeeded from 27 patients 


# Hackensack samples

```{r}
alpha <- asv_alpha_diversity_ag %>% 
  filter(str_detect(oligos_id, 'pool1126|pool1127|pool1129|pool1128')) %>% 
  select(sampleid:shannon, path_pool)  %>% 
  filter(!str_detect(sampleid, 'MB'))
dir.create('../data/hackensack')
alpha %>% write_csv('../data/hackensack/samples_alpha_diversity.csv')

cts <- get_counts_subset(alpha$sampleid) %>% 
  select(asv_key:count_total, count_relative)

taxa <- asv_annotation_blast_ag %>% 
  filter(asv_key %in% cts$asv_key) %>% 
  select(asv_key:blast_pass)

list(alpha = alpha,
     counts = cts,
     taxa = taxa) %>% 
  imap(function(.x, .y){
    write_csv(.x, str_glue('../data/hackensack/hackensack_ASV_{.y}.csv'))
  })



```

```{r}
# downloading the 153 hackensack samples to local 
alpha %>% 
  select(oligos_id, path_pool) %>% 
   mutate(R1 = str_glue('{oligos_id}_R1.fastq.gz'),
         R2 = str_glue('{oligos_id}_R2.fastq.gz')) %>% 
  mutate(p1 = str_glue('{path_pool}/{R1}'),
         p2 = str_glue('{path_pool}/{R2}')) %>% 
   select(p1, p2) %>% 
   gather() %>% 
   mutate(cmd = str_glue('rsync --progress --partial -avz daia1@lilac.mskcc.org:{value} .')) %>% 
   select(cmd) %>% 
   write_csv('../data/hackensack/dl153.sh', col_names = F)
```

```{r}
# find the oligos of the samples from Emily's file
fns <- list.files('../data/hackensack/', pattern = '^pool', full.names = T)


all_oligos <- fns %>% 
  set_names(fns) %>% 
  map( ~ read_tsv(., skip = 2, comment = '#', col_names = F)) %>% 
  bind_rows(.id = 'pool') %>% 
  filter(!str_detect(X4, 'MB')) %>% 
  mutate(sampleid = str_replace_all(X4, '-','\\.'))


# what happened to the missing 163-151 = 12 samples 
missing <- all_oligos %>% 
  filter(!sampleid %in% alpha$sampleid)

missing %>% 
  select(sampleid) %>% 
  write_csv('../data/missing12_hackensack.csv')
```

```{r}
get_counts_subset(c('ENG1080.2.4M'))
```

```{r}
# pool 1137 for Urvi
alpha <- asv_alpha_diversity_ag %>% 
  filter(uploaded_date == '2022-07-07') %>% 
  select(sampleid:shannon)

cts <- get_counts_subset(alpha$sampleid) %>% 
  select(asv_key:count_total, count_relative)

taxa <- asv_annotation_blast_ag %>% 
  filter(asv_key %in% cts$asv_key) %>% 
  select(asv_key:blast_pass)

list(alpha = alpha,
     counts = cts,
     taxa = taxa) %>% 
  imap(function(.x, .y){
    write_csv(.x, str_glue('../data/pool1137_Urvi_ASV_{.y}.csv'))
  })

cts %>% distinct(sampleid)
```
